networking: Move ResizableSemaphore.
