+++ crates/subspace-networking/src/constructor.rs
@@ -14,7 +14,8 @@ use crate::protocols::peer_info::PeerInfoProvider;
 use crate::protocols::request_response::request_response_factory::RequestHandler;
 use crate::protocols::reserved_peers::Config as ReservedPeersConfig;
 use crate::shared::Shared;
-use crate::utils::{strip_peer_id, ResizableSemaphore};
+use crate::utils::rate_limiter::resizable_semaphore::ResizableSemaphore;
+use crate::utils::strip_peer_id;
 use crate::{PeerInfo, PeerInfoConfig};
 use backoff::{ExponentialBackoff, SystemClock};
 use futures::channel::mpsc;
