+++ crates/subspace-networking/src/node.rs
@@ -3,7 +3,8 @@ use crate::protocols::request_response::request_response_factory;
 pub use crate::shared::NewPeerInfo;
 use crate::shared::{Command, CreatedSubscription, Shared};
 use crate::utils::multihash::Multihash;
-use crate::utils::{HandlerFn, ResizableSemaphorePermit};
+use crate::utils::rate_limiter::resizable_semaphore::ResizableSemaphorePermit;
+use crate::utils::HandlerFn;
 use bytes::Bytes;
 use event_listener_primitives::HandlerId;
 use futures::channel::mpsc::SendError;
