+++ crates/subspace-networking/src/utils/rate_limiter.rs
@@ -1,0 +1,3 @@
+pub(crate) mod resizable_semaphore;
+#[cfg(test)]
+mod tests;
